new Kaboom events being implements
scene for handling several events.
go to move between scenes.
start for beginning the game.
and getTag for looping a specify tag.
updates and minors bugs
left and right bumps not being triggered when an obj bot touchs an obj player.
several functions where renamed and moved to obj,props (). check wiki for details.

added obj.Scale for scaling an obj
and  obj.remove () for deleting an obj.